Implement NestJs with GraphQL and TypeORM Server
Added 3 modules for each entity: (1) Users, (2) Roles and (3) Messages.
Each Module has GraphQL resolvers with Queries and Mutations.
Implemented JWT authentication with NestJS guards determining whether a user is logged in or has the appropriate role to access the GraphQL queries and mutations.
Added rough implementation of NestJS Gateway to allow WebSockets from client for GraphQL Subscriptions.
Need to work on Subscriptions and might create a new repo for the NestJS Server.